Relative path fix and name extraction
