+++ public/locales/en/translation.json
@@ -1,5 +1,7 @@
 {
   "bio": {
+    "name": "Ricardo",
+    "surname": "Soto Estévez",
     "p1": "I'm a spanish 25 years old Software Engineer underdegree at University of Oviedo.",
     "p2": "I was born in Noia, Galicia and have being coding half my life. I just enjoy a lot coding, when I'm bored I usually think about creating a new project, probably with whatever I'm learning at that time. For example, this website, I'm learning React and that's how I use the knowledge I'm acquiring, making this really sweets my quarantine.",
     "p3": "Welcome, I'm Ricardo Soto Estévez, software developer.",
