--- src/configs/common.json
@@ -1,4 +1,0 @@
-{
-  "name": "/myreactportfolio",
-  "version": "0.3.0"
-}
