+++ src/containers/Home/home.container.js
@@ -2,7 +2,7 @@ import React from "react";
 import {Col, Image, Row} from "react-bootstrap";
 import {Spacer} from "../../components";
 import {useTranslation} from "react-i18next";
-import {csocial, name} from "../../configs";
+import {csocial} from "../../contexts";
 import me from "./img";
 import "./home.css";
 
@@ -18,7 +18,7 @@ export const HomeComponent = () => {
 				<Col md={12} lg={3}/>
 				<Col md={12} lg={6}>
 					<Image src={me} style={{height: "200px"}} alt="Ricardo Soto Estévez"/>
-					<h1>Ricardo <span className="surname text-primary">Soto Estévez</span></h1>
+					<h1>{t("bio.name")} <span className="surname text-primary">{t("bio.surname")}</span></h1>
 					<hr/>
 					<p>{t("bio.p1")}.</p>
 					<p>{t("bio.p2")}</p>
@@ -37,7 +37,8 @@ export const HomeComponent = () => {
 					<h3>{t("bio.social")}</h3>
 					{csocial.social.map((item) => (
 						<a key={item.key} href={item.to}>
-							<Image src={name + item.icon} style={{height: "50px", margin: "0px 10px"}} alt={item.alt}/>
+							<Image src={process.env.PUBLIC_URL + item.icon} style={{height: "50px", margin: "0px 10px"}}
+								   alt={item.alt}/>
 						</a>
 					))}
 				</Col>
