+++ src/containers/Portfolio/ProjectsComponents/main-project.component.js
@@ -1,7 +1,6 @@
 import React from "react";
 import {Col, Image, Row} from "react-bootstrap";
 import "./projects.css";
-import {name} from "../../../configs/common";
 
 export const MainProjectComponent = () => {
 
@@ -9,7 +8,7 @@ export const MainProjectComponent = () => {
 		<Col md={12} lg={6}>
 			<div className="expositor">
 				<div className="expositor-left">
-					<Image src={name + "/img/logo.svg"} alt={""}
+					<Image src={process.env.PUBLIC_URL + "/img/logo.svg"} alt={""}
 						className="expositor-image"/>
 					<div className="expositor-extra">
 						<h3>Título</h3>
@@ -18,13 +17,16 @@ export const MainProjectComponent = () => {
 						</div>
 						<Row noGutters={true} className="expositor-logos">
 							<Col xs={2} sm={2} md={2} lg={2} xl={2}>
-								<Image src={name + "/img/lang/flag-uk.svg"} alt={""} width={"100%"}/>
+								<Image src={process.env.PUBLIC_URL + "/img/lang/flag-uk.svg"} alt={""}
+									   className={"expositor-logo"} data-toggle="tooltip" title="Hooray!"/>
 							</Col>
 							<Col xs={2} sm={2} md={2} lg={2} xl={2}>
-								<Image src={name + "/img/lang/flag-es.svg"} alt={""} width={"100%"}/>
+								<Image src={process.env.PUBLIC_URL + "/img/lang/flag-es.svg"} alt={""}
+									   className={"expositor-logo"} data-toggle="tooltip" title="Hooray!"/>
 							</Col>
 							<Col xs={2} sm={2} md={2} lg={2} xl={2}>
-								<Image src={name + "/img/lang/flag-gl.svg"} alt={""} width={"100%"}/>
+								<Image src={process.env.PUBLIC_URL + "/img/lang/flag-gl.svg"} alt={""}
+									   className={"expositor-logo"} data-toggle="tooltip" title="Hooray!"/>
 							</Col>
 						</Row>
 					</div>
