+++ src/contexts/index.js
@@ -1,13 +1,9 @@
-import ccommons from "./common";
 import clanguages from "./languages";
 import cnavbar from "./navbar-options.json";
 import csocial from "./social-media";
-import {name} from "./common";
 
 export {
-	ccommons,
 	clanguages,
 	cnavbar,
 	csocial,
-	name
 };
