+++ src/i18n.js
@@ -2,7 +2,6 @@ import i18n from "i18next";
 import Backend from "i18next-xhr-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
-import {name} from "./configs";
 
 i18n
 	.use(Backend)				// loads translations
@@ -13,8 +12,8 @@ i18n
 		debug: true,
 		saveMissing: true,
 		backend: {
-			loadPath: name + "/locales/{{lng}}/{{ns}}.json",
-			addPath: name + "/locales/{{lng}}/{{ns}}.missing.json",
+			loadPath: process.env.PUBLIC_URL + "/locales/{{lng}}/{{ns}}.json",
+			addPath: process.env.PUBLIC_URL + "/locales/{{lng}}/{{ns}}.missing.json",
 			allowMultiLoading: true,
 		},
 		interpolation: {
