+++ src/routes.js
@@ -5,7 +5,6 @@ import {
 } from "react-router-dom";
 import { Home, Portfolio } from "./containers/";
 import { PublicLayout } from "./layouts";
-import { name } from "./configs";
 
 /**
  * Manages the BrowserRouter and so, all the route navigation
@@ -14,12 +13,12 @@ import { name } from "./configs";
  */
 const Routes = () => {
 	return (
-		<Router>
+		<Router basename="/myreactportfolio">
 			<Fragment>
 				{/* Chooses the first route matching the direction and loads it */}
 				<Switch>
-					<PublicLayout component={Home} path={name + "/bio"} exact/> {/* Bio - "/bio" */}
-					<PublicLayout component={Portfolio} path={name + "/portfolio"} exact/> {/* Bio - "/bio" */}
+					<PublicLayout component={Home} path={"/bio"} exact/> {/* Bio - "/bio" */}
+					<PublicLayout component={Portfolio} path={"/portfolio"} exact/> {/* Bio - "/bio" */}
 					<PublicLayout component={Home} path="*"/> {/* All the non existent pages redirects to the Home */}
 				</Switch>
 			</Fragment>
